requirements.txt
fastapi~=0.115.12
uvicorn==0.23.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.4.2
transformers~=4.51.3
torch~=2.7.0
numpy~=2.2.5